pyproject.toml
[project]
name = "educa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django~=5.0.4",
    "django-braces==1.15.0",
    "django-debug-toolbar==4.3.0",
    "django-embed-video>=1.4.10",
    "django-redisboard>=9.0.0",
    "djangorestframework==3.15.1",
    "pillow==10.3.0",
    "pymemcache==4.0.0",
    "redis==5.0.4",
]
